feat/메인페이지 UI작업
